+++ tests/Application/Query/Happening/Webinar/GetWebinarViewQueryHandlerTest.php
@@ -19,7 +19,9 @@ use Proximum\Vimeet\Application\View\Happening\WebinarParticipantView;
 use Proximum\Vimeet\Application\View\Happening\WebinarSpeakerView;
 use Proximum\Vimeet\Application\View\Happening\Webinar\SpeakerWebinarView;
 use Proximum\Vimeet\Application\View\Happening\Webinar\ViewerWebinarView;
+use Proximum\Vimeet\Application\View\Happening\Webinar\WaitingMediaView;
 use Proximum\Vimeet\Domain\Happening\Webinar\IsRecordingAllowed;
+use Proximum\Vimeet\Domain\MimeType\MimeType;
 use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Domain\Model\Happening;
 use Proximum\Vimeet\Domain\Model\HappeningParticipation;
@@ -541,6 +543,8 @@ class GetWebinarViewQueryHandlerTest extends TestCase
         $happening->getBegin()->shouldBeCalled()->willReturn(new \DateTime('2020-03-30 11:00:00'));
         $happening->getEnd()->shouldBeCalled()->willReturn(new \DateTime('2020-03-30 11:45:00'));
         $happening->getWebinarHeaderImage('en')->shouldBeCalled()->willReturn('/path/image.jpg');
+        $happening->getWebinarWaitingMediaFile('en')->shouldBeCalled()->willReturn(null);
+        $happening->getWebinarWaitingMediaType('en')->shouldBeCalled()->willReturn(null);
         $happening->getLiveUrl()->shouldBeCalled()->willReturn('https://www.utube.com/embed/whatever');
         $happening->isSidebarAllowed()->shouldBeCalled()->willReturn(true);
         $happening->isVideoWebinarAndHasLiveUrl()->shouldBeCalled()->willReturn(true);
@@ -576,6 +580,7 @@ class GetWebinarViewQueryHandlerTest extends TestCase
                 $this->dateTime,
                 0,
                 '/path/image.jpg',
+                new WaitingMediaView(null, null),
                 'https://www.utube.com/embed/whatever',
                 true,
                 true,
@@ -740,6 +745,8 @@ class GetWebinarViewQueryHandlerTest extends TestCase
         $happening->getBegin()->shouldBeCalled()->willReturn(new \DateTime('2020-03-30 11:55:00'));
         $happening->getEnd()->shouldBeCalled()->willReturn(new \DateTime('2020-03-30 12:15:00'));
         $happening->getWebinarHeaderImage('en')->shouldBeCalled()->willReturn('/path/image.jpg');
+        $happening->getWebinarWaitingMediaFile('en')->shouldBeCalled()->willReturn('/path/to/video.mp4');
+        $happening->getWebinarWaitingMediaType('en')->shouldBeCalled()->willReturn(MimeType::FORMAT_VIDEO);
         $happening->getLiveUrl()->willReturn(null);
         $this->questionRepository->getMessagesCountDuringHappening($happening->reveal())->shouldBeCalled()->willReturn(21);
         $happening->isSidebarAllowed()->shouldBeCalled()->willReturn(true);
@@ -794,6 +801,7 @@ class GetWebinarViewQueryHandlerTest extends TestCase
                 $this->dateTime,
                 900,
                 '/path/image.jpg',
+                new WaitingMediaView('/path/to/video.mp4', MimeType::FORMAT_VIDEO),
                 null,
                 true,
                 false,