+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - MV285 - Custom button display on mobile
 
+### Added
+- VIMEET-2252 - Reactivate account
+
+## [2.72.1] - 2020-12-02
+### Fixed
+- MV-286 - Fix disconnection when alert is not closed
+
+## [2.72.0] - 2020-12-02
+### Added
+- VIMEET-2185 - Add waiting image/video to webinar
+- VIMEET-2248 - Help tip networking
+
 ### Updated
 - VIMEET-2121 - Change visio test process, fix regressions due to changes in VideoConference